Bug fix to allow deletion of modules w/o state folder because they have never been initialized

+++ sensorhub-core/src/main/java/org/sensorhub/impl/module/DefaultModuleStateManager.java
@@ -215,7 +215,8 @@ public class DefaultModuleStateManager implements IModuleStateManager
     {
         try
         {
-            FileUtils.deleteRecursively(folder);
+            if (folder.exists())
+                FileUtils.deleteRecursively(folder);
         }
         catch (IOException e)
         {
